inttests/requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
mockserver-client==0.0.5
requests==2.19.1
urllib3==1.23
python-dateutil===2.7.3
pytest==3.7.0